refactor: (main) Route Group 생성 및 Header 통합 관리
- (main)/layout.tsx에서 Header를 단일 관리 (container-1920 pt-6.25 적용)
- home, about-us, profile, performance-list, performance-detail을 (main)/로 편입
- 각 page.tsx에서 중복 등록된 Header 및 mt-6.25 래퍼 제거
