+++ src/app/(main)/about-us/_components/about-us-hero-section.tsx
