+++ src/app/(main)/about-us/_components/index.ts
