+++ src/app/(main)/performance-detail/[performanceId]/_components/footer.tsx
