+++ src/app/(main)/performance-detail/[performanceId]/_components/index.ts
