+++ src/app/(main)/performance-detail/[performanceId]/_components/performance-info.tsx
