+++ src/app/(main)/performance-list/_components/hero.tsx
